Update index.html
Your corrected CSS code:

```css
background: linear-gradient(to bottom, red, blue, pink);
```

This will create a linear gradient background that transitions from red at the top to blue in the middle and then to pink at the bottom. The `to bottom` specifies the direction of the gradient, indicating that it should go from the top of the element to the bottom. Feel free to modify the colors or add more color stops if you want to customize the gradient further.